circular dependency - code re-ordering, cleanupreduant code. New files are created to achieve code readability
