+++ tests/test_fin_utils.py
@@ -29,7 +29,7 @@ from tuna.miopen.db.miopen_tables import ConvolutionConfig, ConvolutionJob, Tens
 from multiprocessing import Value, Lock, Queue
 from tuna.utils.metadata import LOG_TIMEOUT
 from tuna.miopen.db.tables import MIOpenDBTables, ConfigType
-from tuna.miopen.utils.session_utils import get_session_t
+from tuna.miopen.db.session import Session
 
 
 def test_fin_utils():
@@ -39,7 +39,7 @@ def test_fin_utils():
   my_job.valid = 1
   my_job.config = 1
   dbt = MIOpenDBTables(session=1, config_type=ConfigType.convolution)
-  dbt.session = get_session_t()
+  dbt.session = Session()
   dbt.session.id = 1
   dbt.session.arch = 'gfx908'
   dbt.session.num_cu = 120
